수업
페이징...하고 또 뭐 했는데 기억안남

+++ webPrj99semi/WebContent/views/notice/noticeDetail.jsp
@@ -58,10 +58,14 @@
 
         <div>
             <a href="<%=contextPath%>/notice/list" class="btn btn-sm btn-primary">목록이동</a>
+            
+            <%if(loginMember!=null && vo.getWriter().equals(loginMember.getName())){%>
 
-            <!-- 현재 로그인한 사용자가 해당 글을 쓴 본인일 경우 -->
-            <a href="<%=contextPath%>/notice/edit?num=<%=vo.getNo()%>" class="btn btn-sm btn-warning">수정하기</a>
-            <a href="<%=contextPath%>/notice/delete?num=<%=vo.getNo()%>" class="btn btn-sm btn-danger">삭제하기</a>
+	            <!-- 현재 로그인한 사용자가 해당 글을 쓴 본인일 경우 -->
+	            <a href="<%=contextPath%>/notice/edit?num=<%=vo.getNo()%>" class="btn btn-sm btn-warning">수정하기</a>
+	            <a href="<%=contextPath%>/notice/delete?num=<%=vo.getNo()%>" class="btn btn-sm btn-danger">삭제하기</a>
+            
+            <%} %>
         </div>
 
     </div>
